+++ packages/edit-site/src/components/page-patterns/fields.js
@@ -9,10 +9,7 @@ import clsx from 'clsx';
 import { __experimentalHStack as HStack } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
 import { useState, useMemo, useId } from '@wordpress/element';
-import {
-	BlockPreview,
-	privateApis as blockEditorPrivateApis,
-} from '@wordpress/block-editor';
+import { BlockPreview } from '@wordpress/block-editor';
 import { Icon } from '@wordpress/icons';
 import { parse } from '@wordpress/blocks';
 
@@ -24,16 +21,14 @@ import {
 	PATTERN_SYNC_TYPES,
 	OPERATOR_IS,
 } from '../../utils/constants';
-import { unlock } from '../../lock-unlock';
 import { useAddedBy } from '../page-templates/hooks';
-
-const { useGlobalStyle } = unlock( blockEditorPrivateApis );
+import { useStyle } from '../global-styles';
 
 function PreviewField( { item } ) {
 	const descriptionId = useId();
 	const description = item.description || item?.excerpt?.raw;
 	const isTemplatePart = item.type === TEMPLATE_PART_POST_TYPE;
-	const [ backgroundColor ] = useGlobalStyle( 'color.background' );
+	const backgroundColor = useStyle( 'color.background' );
 	const blocks = useMemo( () => {
 		return (
 			item.blocks ??